requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
django-filter==24.1
django-cors-headers==4.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3

django-cors-headers==4.3.1
